deta_test/cmprxn_test1/requirements.txt
fastapi
pillow
numpy
matplotlib
requests